% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANCRTAdjust.R
\docType{package}
\name{ANCRTAdjust}
\alias{ANCRTAdjust}
\alias{ANCRTAdjust-package}
\title{ANCRTAdjust: A package to adjust ANC-RT HIV prevalence estimates}
\description{
This package provides a variety of functions to prepare ANC-RT data for analysis and to
ultimately adjust HIV prevalences for bias caused by missing and invalid data.
}
\author{
Mathieu Maheu-Giroux \email{mathieu.maheu-giroux@mcgill.ca}
}
