% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Clean ANC-RT Data}
\usage{
data_clean(data, total_age_cat = NULL)
}
\arguments{
\item{data}{A country-specific ANC-RT database. The function \link[ANCRTAdjust]{name_var} should have been run on the data to properly
prepare the data for use here.
The following variables (with these variable names) must be included in \code{data}:
 \itemize{
  \item \code{time}: The time period that the data was collected.
  \item \code{faciluid}:  The unique facility identifier.
  \item \code{n_clients}: The number of women from the specified facility, during the specified time period, that attended their first ANC visit.
  \item \code{n_status}:  The number of women from the specified facility, during the specified time period, that had their HIV status ascertained at their first ANC visit, either by testing or through previous knowledge.
  \item \code{testpos}: The number of women from the specified facility, during the specified time period, that tested positive for HIV at their first ANC visit.
  \item \code{knownpos}: The number of women from the specified facility, during the specified time period, that already knew that they were HIV-positive at their first ANC visit.
  \item \code{testneg}: The number of women from the specified facility, during the specified time period, that tested negative for HIV at their first ANC visit.
  \item \code{totpos}: The number of women from the specific facility, during the specified time period, that were HIV-positive at their first ANC visit. 
  \item \code{age}: Age category of pregnant women (optional).
 }}

\item{total_age_cat}{The value of the \code{age} variable indicating all age categories combined (only necessary if age disaggregated data is included).}
}
\value{
A cleaned ANC-RT database with additional variables:
 \itemize{
  \item \code{testneg_c}: Cleaned \code{testneg} (according to description)
  \item \code{testpos_c}: Cleaned \code{testpos} (according to description)
  \item \code{knownpos_c}: Cleaned \code{knownpos} (according to description)
  \item \code{n_status_c}: Cleaned \code{n_status} (according to description)
  \item \code{n_status_c.impute}:  Adjusted \code{n_status_c} using the impute adjustment option for multiple testing
  \item \code{n_status_c.remove}:  Adjusted \code{n_status_c} using the remove adjustment option for multiple testing
  \item \code{n_status_c.setmax}:  Adjusted \code{n_status_c} using the set to maximum adjustment option for multiple testing
  \item \code{totpos_c}: Cleaned \code{totpos} (according to description)
  \item \code{totpos_c.impute}: Adjusted \code{totpos_c} if the impute adjustment option for multiple testing is used
  \item \code{totpos_c.remove}: Adjusted \code{totpos_c} if the remove adjustment option for multiple testing is used
  \item \code{totpos_c.setmax}: Adjusted \code{totpos_c} if the set to maximum adjustment option for multiple testing is used
 }
}
\description{
Cleans ANC-RT data so that it can be used with other functions in the \link[ANCRTAdjust] package.
}
\details{
This function has been developed to clean common issues asoociated with ANC-RT data.  The following tasks
are accomplished:
 \itemize{
   \item Age disaggregated data is removed (if applicable).
   \item \code{testneg_c} is set to \code{n_status} - \code{knownpos} - \code{testpos} or \code{n_status} - \code{totpos} if \code{testneg} is missing.  \code{testneg_c} is set to 0 if negative.
   \item \code{knownpos_c} is set to \code{n_status} - \code{testneg_c} - \code{testpos} or \code{totpos} - \code{testpos} if \code{knownpos} is missing.  \code{knownpos_c} is set to 0 if negative.
   \item \code{testpos_c} is set to \code{n_status} - \code{testneg_c} - \code{knownpos_c} or \code{totpos} - \code{knownpos} if \code{testpos} is missing.  \code{testpos_c} is set to 0 if negative.
   \item \code{n_status_c} is set to \code{testpos_c} + \code{testneg_c} + \code{knownpos_c} when all of these variables are available.  If either of the three variables is
   missing, \code{n_status_c} = \code{n_status}.
   \item \code{totpos_c} is set to \code{knownpos_c} + \code{testpos_c} if both of these variables are available. If either \code{testpos_c} or \code{knownpos_c} are 
   missing but \code{totpos} is available, \code{totpos_c} is set to \code{totpos}.  If \code{totpos_c} is missing but \code{n_status_c} and \code{testneg_c} are available, \code{totpos_c} = \code{n_status_c} - \code{testneg_c}. 
   \item The three suggested \code{n_status_c} adjustments are created in the case that \code{n_status_c} > \code{n_clients}:
   \itemize{
     \item \code{n_status_c.impute} uses the impute adjustment option
     \item \code{n_status_c.remove} uses the remove adjustment option
     \item \code{n_status_c.setmax} uses the set to maximum adjustment option
     }
   \item \code{totpos_c} is adjusted if \code{totpos_c} > \code{n_status_c}:
     \itemize{
       \item \code{totpos_c}: \code{totpos_c} is set to missing when \code{totpos_c} > \code{n_status_c}
       \item \code{totpos_c.impute}: \code{totpos_c} is set to missing when \code{totpos_c} > \code{n_status_c.impute}
       \item \code{totpos_c.remove}: \code{totpos_c} is set to missing when \code{totpos_c} > \code{n_status_c.remove}
       \item \code{totpos_c.setmax}: \code{totpos_c} is set to missing when \code{totpos_c} > \code{n_status_c.setmax}
       }
   \item The data is checked for duplicates (i.e. more than one observation exists with the same \code{faciluid} and \code{time}).
  }
}
\author{
Mathieu Maheu-Giroux

Brittany Blouin
}
