% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_outliers.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Flag outlier observations}
\usage{
flag_outliers(data, sd_for_outlier = 2, flag_by = "facility",
  result = "outliers")
}
\arguments{
\item{data}{The ANC-RT dataset.  The functions \link[ANCRTAdjust]{check_data}, \link[ANCRTAdjust]{data_clean} and \link[ANCRTAdjust]{mt_adjust} should have been run on the data to properly
prepare the data for use here.  The dataset must have the following variables:
 \itemize{
  \item \code{faciluid}: Facility ID.
  \item \code{time}: The time period over which the data was collected.
  \item \code{n_clients}: The number of women from the specified facility, during the specified time period, that attended their first ANC visit.
  \item \code{n_status_c}:  The cleaned number of women from the specified facility, during the specified time period, that had their HIV status ascertained at their first ANC visit, either by testing or through previous knowledge
  (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{testpos_c}: The cleaned number of women from the specified facility, during the specified time period, that tested positive for HIV at their first ANC visit
  (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{knownpos_c}: The cleaned number of women from the specified facility, during the specified time period, that already knew that they were HIV-positive at their first ANC visit
  (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{testneg_c}: The cleaned number of women from the specified facility, during the specified time period, that tested negative for HIV at their first ANC visit 
  (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{totpos_c}: The cleaned total number of positive HIV cases (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{prv}:  The HIV prevalence from the specified facility at the specified time period (generated using the \link[ANCRTAdjust]{mt_adjust} function).
  \item \code{cov}:  The HIV testing coverage from the specified facility at the specified time period (generated using the \link[ANCRTAdjust]{mt_adjust} function).
  \item \code{snu1}: The subnational unit 1 (only required if results are to be flagged by snu1).
   }}

\item{sd_for_outlier}{Standard deviation used to flag outliers (default is 2).}

\item{flag_by}{Options include:
\itemize{
  \item "\code{facility}" compares each observation's value to their facility's mean value and flags the observations that 
  are greater than or less than 2 standard deviations from the facility mean.
  \item "\code{snu1}" compares each observation's value to their sub national unit 1's mean value and flags the observations that 
  are greater than or less than 2 standard deviations from the snu1 mean.
  \item "\code{country}" compares each observation's value to their country's mean value and flags the observations that 
  are greater than or less than 2 standard deviations from the country mean.
  }}

\item{result}{Options include:
\itemize{
   \item "\code{outliers}" returns a dataset including the observations that are considered to have an outlier value for any of:
   \code{n_clients}, \code{n_status_c}, \code{testpos_c}, \code{testneg_c}, \code{knownpos_c}, \code{totpos_c}, \code{prv} or \code{cov}. The values 
   for each of the eight variables are only reported if they are considered an outlier. If they are not considered an outlier, they are reported 
   as "NA". For identification purposes \code{faciluid} and \code{time} are also included.
   \item "\code{data}" returns the complete dataset (that was originally input into the function) with the following additional variables:
     \itemize{
        \item \code{flag_n_clients}: A value of 1 indicates that the \code{n_clients} value is considered an outlier and a value of 0 indicates that it is not considered an outlier.
        \item \code{flag_n_status_c}: A value of 1 indicates that the \code{n_status_c} value is considered an outlier and a value of 0 indicates that it is not considered an outlier.
        \item \code{flag_testpos_c}: A value of 1 indicates that the \code{testpos_c} value is considered an outlier and a value of 0 indicates that it is not considered an outlier.
        \item \code{flag_testneg_c}: A value of 1 indicates that the \code{testneg_c} value is considered an outlier and a value of 0 indicates that it is not considered an outlier.
        \item \code{flag_knownpos_c}: A value of 1 indicates that the \code{knownpos_c} value is considered an outlier and a value of 0 indicates that it is not considered an outlier.
        \item \code{flag_totpos_c}: A value of 1 indicates that the \code{totpos_c} value is considered an outlier and a value of 0 indicates that it is not considered an outlier.
        \item \code{flag_prv}: A value of 1 indicates that the \code{prv} value is considered an outlier and a value of 0 indicates that it is not considered an outlier.
        \item \code{flag_cov}: A value of 1 indicates that the \code{cov} value is considered an outlier and a value of 0 indicates that it is not considered an outlier.
        }
 }}
}
\value{
A dataset including either the flagged observations only or the full, original dataset with additional variables indicating flagged observations, according to user inputs.
}
\description{
Flags observations that are considered outliers
}
\details{
This function has been developed to flag outlier observations for the following variables: \code{n_clients}, \code{n_status_c}, \code{testpos_c}, \code{testneg_c}, \code{knownpos_c}, 
\code{totpos_c}, \code{prv} and \code{cov}.  Outliers are defined as 2 standard deviations greater than or less than the mean value.
}
\author{
Mathieu Maheu-Giroux

Brittany Blouin
}
