% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impcov_adjust_simple.R
\name{impcov_adjust_simple}
\alias{impcov_adjust_simple}
\title{Adjustment for imperfect HIV testing coverage}
\usage{
impcov_adjust_simple(hiv_prv, hiv_cov)
}
\arguments{
\item{hiv_prv}{The HIV prevalence adjusted for all previous adjustments (e.g. data cleaning, adjustment 
for multiple testing and/or adjustment for missing reporting periods).  This MUST be input as a proportion, between 0-1.}

\item{hiv_cov}{The HIV testing coverage adjusted for all previous adjustments (e.g. data cleaning, adjustment 
for multiple testing and/or adjustment for missing reporting periods).  This MUST be input as a proportion, between 0-1.}
}
\value{
The HIV prevalence adjusted for imperfect HIV testing coverage.
}
\description{
Adjusts the HIV prevalence for imperfect HIV testing coverage
}
\details{
This function was designed to adjust HIV prevalence when HIV testing coverage is less than 100\%.
}
\author{
Mathieu Maheu-Giroux

Brittany Blouin
}
