% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imperfectcov_adjust.R
\name{imperfectcov_adjust}
\alias{imperfectcov_adjust}
\title{Imperfect coverage adjustment}
\usage{
imperfectcov_adjust(data, byperiod = "FALSE", bysnu1 = "FALSE",
  byyear = "FALSE")
}
\arguments{
\item{data}{The ANC-RT dataset.  Ideally the functions \link[ANCRTAdjust]{data_clean} and \link[ANCRTAdjust]{mt_adjust} have been run on the data to properly
prepare the data for use here.  The dataset must have the following variables:
 \itemize{
  \item \code{ID}: Facility ID (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{snu1}: The subnational unit 1.
  \item \code{Time}: The quarter that the data was collected (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{Year}: The year that the data was collected.
  \item \code{testpos}: The number of women from the specified facility, during the specified time period, that tested positive for HIV at their first ANC visit.
  \item \code{knownpos}: The number of women from the specified facility, during the specified time period, that already knew that they were HIV-positive at their first ANC visit.
  \item \code{n_status}:  The number of women from the specified facility, during the specified time period, that had their HIV status ascertained at their first ANC visit, either by testing or through previous knowledge.
  \item \code{TotPos}: Total number of positive HIV cases (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{n_stat}: Cleaned \code{n_status} (generated using the \link[ANCRTAdjust]{data_clean} function).
  \item \code{n_clients}: The number of women from the specified facility, during the specified time period, that attended their first ANC visit.
 }}

\item{byperiod}{"TRUE" indicates that the results be stratified by reporting period.  "FALSE" indicates that the results not be stratified.}

\item{bysnu1}{"TRUE" indicates that the results be stratified by snu1.  "FALSE" indicates that the results not be stratified.}

\item{byyear}{"TRUE" indicates that the results be stratified by year.  "FALSE" indicates that the results not be stratified.}
}
\value{
A table (in dataframe format) indicating the results, stratified according to user inputs.  The results table includes the following columns:
 \itemize{
  \item{snu1}: The subnational unit 1.  "All" refers to results not stratified by snu1.
  \item{Time}: The time frame.  When results are stratified by year and reporting period, the digits "99" are suffixed to the year (i.e. "201599" refers to the yearly result for 2015).
  "9999" refers to the results not stratified by time (i.e. not stratified by period or year).
  \item{HIV_fulladjust}: The HIV prevalence adjusted for imperfect testing coverage and all previous adjustments (i.e. data cleaning and adjustment for multiple testing, if performed).
  \item{RR_observed}: The relative bias comparing the HIV prevalence at the observed HIV testing coverage to the predicted HIV prevalence when HIV testing coverage is set to 100\%.
  \item{RR_90}: The relative bias comparing the predicted HIV prevalence when HIV testing coverage is set to 90\% to the predicted HIV prevalence when HIV testing coverage is set to 100\%.
  \item{HIV_raw}: The HIV prevalence using the raw data.
  \item{HIV_cleaned}: The HIV prevalence using the cleaned data (and following adjustment for multiple testing, if performed).
 }
}
\description{
Adjusts HIV prevalence from ANC-RT data for imperfect HIV testing coverage
}
\details{
This function has been developed to adjust HIV prevalence from ANC-RT data for imperfect (i.e. less than 100\%) HIV testing coverage.
The adjustment model is a logisitic regression model that models HIV prevalence conditional on HIV testing coverage (relationship modeled with
a natural sline with one knot at 33\%), time (relationship modeled as a fixed effect) and facility (relationship modeled as a fixed effect).  
The adjusted HIV prevalence can be reported stratified by reporting period, year and/or subnational unit 1, according to user inputs.
}
\author{
Mathieu Maheu-Giroux

Brittany Blouin
}
