% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_var.R
\name{name_var}
\alias{name_var}
\title{Prepares ANC-RT data}
\usage{
name_var(data = NULL, faciluid = NULL, time = NULL,
  n_clients = NULL, n_status = NULL, knownpos = NULL,
  testpos = NULL, testneg = NULL, totpos = NULL, age = NULL,
  snu1 = NULL, Year = NULL)
}
\arguments{
\item{data}{A country-specific ANC-RT database.  The following variables must be included in \code{data}:
\itemize{
 \item The time period that the data was collected.
 \item The unique facility identifier.
 \item The number of women from the specified facility, during the specified time period, that attended their first ANC visit.
 }
Optional variables that can be included for re-naming in \code{data} include: 
 \itemize{
  \item The number of women from the specified facility, during the specified time period, that had their HIV status ascertained at their first ANC visit, either by testing or through previous knowledge.
  \item The number of women from the specified facility, during the specified time period, that tested negative for HIV at their first ANC visit.
  \item The number of women from the specified facility, during the specified time period, that tested positive for HIV at their first ANC visit.
  \item The number of women from the specified facility, during the specified time period, that already knew that they were HIV-positive at their first ANC visit.
  \item Age category of pregnant women.
  \item The number of women from the specified facility, during the specified time period, that were HIV-positive at their first ANC visit.
  \item The sub-national unit 1.
  \item The calendar year from which the data was collected.
  }}

\item{faciluid}{The variable name in \code{data} for the unique facility identifier}

\item{time}{The variable name in \code{data} for the time period that the data was collected}

\item{n_clients}{The variable name in \code{data} for the number of women from the specified facility, during the specified time period, that attended their first ANC visit.}

\item{n_status}{The variable name in \code{data} for the number of women from the specified facility, during the specified time period, that had their HIV status ascertained at their first ANC visit, either by testing or through previous knowledge
(if available).  If this variable is not specified, it will be created and marked as missing for all observations.}

\item{knownpos}{The variable name in \code{data} for the number of women from the specified facility, during the specified time period, that already knew that they were HIV-positive at their first ANC visit (if available).
If this variable is not specified, it will be created and marked as missing for all observations.}

\item{testpos}{The variable name in \code{data} for the number of women from the specified facility, during the specified time period, that tested positive for HIV at their first ANC visit (if available).
If this variable is not specified, it will be created and marked as missing for all observations.}

\item{testneg}{The variable name in \code{data} for the number of women from the specified facility, during the specified time period, that tested negative for HIV at their first ANC visit (if available).
If this variable is not specified, it will be created and marked as missing for all observations.}

\item{totpos}{The variable name in \code{data} for the number of women from the specified facility, during the specified time period, that were HIV-positive (if available).  If this variable name is not specified it will 
be automatically created by summing \code{testpos} and \code{knownpos}.}

\item{age}{The variable name in \code{data} for the age category of pregnant women (if available).}

\item{snu1}{The variable name in \code{data} for the sub-national unit 1 (if available).}

\item{Year}{The variable name in \code{data} for the sub-national unit 1 (if available).}
}
\value{
The ANC-RT dataset with variables created and/or re-named, if necessary, ensuring that subsequent package functions will work.
}
\description{
Prepares the ANC-RT dataset to ensure that subsequent functions in the \code{ANCRTAdjust} package work properly.
}
\details{
This function has been developed to prepare the ANC-RT dataset for use within the \code{ANCRTAdjust} package.  
The function ensures that all necessary variables are included in the dataset and renames the variables to ensure that they conform to the 
naming conventions necessary for use in this package.
}
\author{
Mathieu Maheu-Giroux

Brittany Blouin
}
