% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rawadjusted.R
\name{plot_rawadjusted}
\alias{plot_rawadjusted}
\title{Plot of HIV prevalence over time}
\usage{
plot_rawadjusted(Data, snu1 = "All", time.unit = "PERIOD",
  HIV_raw = "TRUE", y.lim = 40)
}
\arguments{
\item{Data}{The dataframe output by \link[ANCRTAdjust]{imperfectcov_adjust}.  The following variables must be present:
\itemize{
 \item \code{snu1}: The subnational unit 1.  If plots are to be generated for the different snu1s, the parameter bysnu1 = "TRUE" must 
 have been specified when running \link[ANCRTAdjust]{imperfectcov_adjust}.
 \item \code{Time}: The year and/or quarter that the data was collected.  If plots are to be generated by reporting period, the parameter 
 byperiod = "TRUE" must have been specified when running \link[ANCRTAdjust]{imperfectcov_adjust}.  If plots are to be generated by year, the parameter
 byyear = "TRUE" must have been specified when running \link[ANCRTAdjust]{imperfectcov_adjust}.
 \item \code{HIV_fulladjust}: The HIV prevalence adjusted for imperfect testing coverage and all previous adjustments (i.e. data cleaning and adjustment 
 for multiple testing, if performed).  Generated using the \link[ANCRTAdjust]{imperfectcov_adjust} function.
 \item{HIV_raw}: The HIV prevalence using the raw data. Generated using the \link[ANCRTAdjust]{imperfectcov_adjust} function.
}}

\item{snu1}{The snu1 for which the plot is generated.  "All" indicates that the plot include the entire country data (i.e. not snu1-specific). Default = "All".}

\item{time.unit}{"PERIOD" indicates that the results be plotted by reporting period.  "YEAR" indicates that the results be plotted by year. Default = "PERIOD".}

\item{HIV_raw}{"TRUE" indicates that both the fully adjusted and raw HIV prevalences be plotted.  "FALSE" indicates that only the fully
adjusted HIV prevalences be plotted. Default = "TRUE".}

\item{y.lim}{The y-axis upper limit. Default = 40.}
}
\value{
A plot of HIV prevalence over time.
}
\description{
Plots the raw and adjusted HIV prevalence over time
}
\details{
This function has been developed to plot HIV prevalence over time.  Both the raw and fully adjusted (obtained using the \link[ANCRTAdjust]{data_clean}, \link[ANCRTAdjust]{mt_adjust}
and \link[ANCRTAdjust]{imperfectcov_adjust} functions) HIV prevalence can be plotted and compared or just the fully adjusted prevalence can be plotted.  The time unit can be set to 
either yearly or by reporting period.
}
\author{
Mathieu Maheu-Giroux

Brittany Blouin
}
