% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.cluster.R
\name{recode.cluster}
\alias{recode.cluster}
\title{Recode cluster}
\usage{
recode.cluster(Var)
}
\arguments{
\item{Var}{The variable that represents the cluster}
}
\value{
The variable \code{ID} which represents cluster IDs, sequentially ordered numbers.
}
\description{
Recodes cluster variables to be a sequential number
}
\details{
This function recodes variables that represent cluster IDs to be sequentially numbered which facilitates
cluster analyses.
}
\author{
Mathieu Maheu-Giroux
}
